fix: footer, add: basic structure productDetail, add: subheader
